requirements.txt
streamlit 
fastai
fastbook
Pillow==8.2.0
gdown 
